refactor: move demographic merge models to commn/model package
