+++ src/main/java/ca/openosp/openo/commn/model/CaseMgmtIssueNotes.java
@@ -3,7 +3,7 @@
  * Copyright (c) 2001-2002. Department of Family Medicine, McMaster University. All Rights Reserved.
  * This software is published under the GPL GNU General Public License.
  */
-package ca.openosp.openo.demographic.merge.model;
+package ca.openosp.openo.commn.model;
 
 import javax.persistence.*;
